pyproject.toml
[tool.poetry]
name = "jinjat"
version = "0.9.8"
description = "A dbt server and suite of optional developer tools to make developing with dbt delightful."
authors = ["z3z1ma <butler.alex2010@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
keywords = ["dbt", "server", "streamlit", "git", "diff"]
documentation = "https://github.com/jinjat-data/jinjat"
repository = "https://github.com/jinjat-data/jinjat"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
click = ">7"
dbt-core = "=1.3.0"
watchdog = ">=2.2.1"
"ruamel.yaml" = ">=0.17"
rich = ">=10"
pydantic = "^1.9.1"
GitPython = "^3.1.27"
bottle = "^0.12.23"
orjson = "^3.8.0"
fastapi = "^0.85.0"
openapi-schema-pydantic = "^1.2.4"
deepmerge = "^1.1.0"
jsonschema = "^3.0"
fastapi-crudrouter = "^0.8.5"
uvicorn = { extras = ["standard"], version = "^0.18.3" }
jinja2-simple-tags = "^0.4.0"
jsonref = "^1.1.0"
cookiecutter = "^2.1.1"
# Streamlit Playground Dependencies
streamlit = { version = ">=1.0.0", optional = true }
streamlit-ace = { version = ">=0.1.0", optional = true }
graphviz = { version = ">=0.17", optional = true }
pydot = { version = ">=1.4.2", optional = true }
streamlit-agraph = { version = ">=0.0.35", optional = true }
streamlit-pandas-profiling = { version = ">=0.1.3", optional = true }
streamlit-aggrid = { version = ">=0.2.2", optional = true }
scipy = { version = "^1.3.1", optional = true }
feedparser = { version = "^6.0.10", optional = true }
# Testing duckdb
duckcli = { version = "^0.2.1", optional = true }
dbt-duckdb = { version = "^1.2.0", optional = true }
jmespath = "^1.0.1"
pandas = "^1.5.3"

[tool.poetry.dev-dependencies]
black = ">=21.9b0"
mypy = ">=0.910"
pytest = ">=6.2.5"
coverage = ">=5.5"
pylint = ">=2.11.1"
nox = ">=2021.6.12"
pytest-mock = ">=3.6.1"
viztracer = "^0.15.3"

[tool.poetry.extras]
duckdb = ["dbt-duckdb", "duckcli"]
playground = [
    "streamlit",
    "streamlit-ace",
    "graphviz",
    "pydot",
    "streamlit-agraph",
    "streamlit-pandas-profiling",
    "streamlit-aggrid",
    "scipy",
    "feedparser",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
line_length = 100
multi_line_output = 3
include_trailing_comma = true

[tool.poetry.scripts]
jinjat = 'jinjat.main:cli'
jinjat-proxy = 'jinjat.server:cli'